# app/openapi.gen.go
// Package app provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.9.1 DO NOT EDIT.
package app

// ActiveUsersResponse defines model for ActiveUsersResponse.
type ActiveUsersResponse struct {
	Count int `json:"count"`
}

// CreateUserRequest defines model for CreateUserRequest.
type CreateUserRequest struct {
	Name     string `json:"name"`
	Password string `json:"password"`
}

// CreateUserResponse defines model for CreateUserResponse.
type CreateUserResponse struct {
	Id   string `json:"id"`
	Name string `json:"name"`
}

// LoginUserRequest defines model for LoginUserRequest.
type LoginUserRequest struct {
	// The user name for login
	Name string `json:"name"`

	// The password for login in clear text
	Password string `json:"password"`
}

// LoginUserResonse defines model for LoginUserResonse.
type LoginUserResonse struct {
	// A url for websoket API with a one-time token for starting chat
	Url string `json:"url"`
}

// StartChatParams defines parameters for StartChat.
type StartChatParams struct {
	// One time token for a loged user
	Token string `json:"token"`
}

// CreateUserJSONBody defines parameters for CreateUser.
type CreateUserJSONBody CreateUserRequest

// LoginUserJSONBody defines parameters for LoginUser.
type LoginUserJSONBody LoginUserRequest

// CreateUserJSONRequestBody defines body for CreateUser for application/json ContentType.
type CreateUserJSONRequestBody CreateUserJSONBody

// LoginUserJSONRequestBody defines body for LoginUser for application/json ContentType.
type LoginUserJSONRequestBody LoginUserJSONBody